+++ lib/src/sdk/invite.dart
@@ -2,7 +2,7 @@ part of 'sdk.dart';
 
 class InviteData {
   final String id;
-  final int guild;
+  final Int64 guild;
   final int usesCount;
 
   InviteData(this.id, this.guild, this.usesCount);
@@ -11,7 +11,7 @@ class InviteData {
 class Invite {
   final Server server;
   final String id;
-  final int guild;
+  final Int64 guild;
   final int usesCount;
 
   Invite(this.server, this.id, this.guild, this.usesCount);
